Extracting `NormalizeIdentifierTest` assets to a `Model` namespace
